// constants/enums/ReportType.ts
export enum ReportType {
  PAYMENT_ISSUE = "PAYMENT_ISSUE",
  RESIDENCE_ISSUE = "RESIDENCE_ISSUE",
  HOST_ISSUE = "HOST_ISSUE",
  OTHER = "OTHER",
}
export enum ReportStatus {
  PENDING = "PENDING",
  INVESTIGATING = "INVESTIGATING",
  RESOLVED = "RESOLVED",
  CLOSE = "CLOSE",
}
export enum ReportSeverity {
  LOW = "LOW",
  MEDIUM = "MEDIUM",
  HIGH = "HIGH",
  URGENT = "URGENT",
}